refractor: type safe, improve io

+++ src/runtime/os/os.hpp
@@ -15,7 +15,7 @@ import <vector>;
 NAMESPACE_BEGIN(os)
 NAMESPACE_BEGIN(detail)
 // @note get unix timestamp (unit: ns)
-auto time() -> usize;
+auto time() -> i64;
 // @note find file in path. example: "src/*.cpp" "src/**.cpp".
 auto find_files(const std::string_view path) -> std::vector<std::string>;
